refactor(ExtractService): 추출 기능 클래스 분할
한 클래스에 많은 로직이 담겨 있어 공통 로직 메서드를 분리했습니다.
